Decompose Backend file and classes into smaller classes & files
